refactor(admin): extract the UI Theme panel from ConfigTab
Last of the three independently-saving panels (plan 013, phase 1). Moves the
UI Theme & Branding panel — handleSaveTheme, handleLogoFile, handleIconFile,
applyThemeToDOM, MAX_LOGO_BYTES and the ten theme state hooks — into
components/admin/config/ThemePanel.tsx. The panel JSX is byte-identical to
what it replaced (verified by diff).

The theme GET moves with it, so the panel now fetches its own values instead
of ConfigTab fetching them on mount. The system config still seeds the colour
and radius (`initialColor` / `initialRadius`) exactly as before, and the theme
endpoint still overrides them — it is the authoritative source and the only
one carrying the org name, logo and icon. Two consequences, both benign:
the theme request now starts once the config has loaded rather than in
parallel with it, and it is no longer issued when the config load fails
(the panel is unreachable in that case anyway, and applyThemeToDOM was never
called on that path).

applyThemeToDOM had exactly one caller, handleSaveTheme, so it moved with it
and the save → apply → branding.refresh() order is unchanged.

Two props. ConfigTab.tsx: 1,518 → 1,243 lines — 2,875 down to 1,243 across
phase 1, with handleSaveConfig and all six sticky-save panels still in it
(phase 2). All 11 ConfigTab characterization tests still pass unchanged, plus
4 new ThemePanel tests. 49 files / 374 tests, coverage gate green;
ConfigTab chunk 87,775 → 87,809 bytes.
